Add batch operations scripts
Added a script to batch generate headers from a list of locales.

Added a script to update codepoints.yaml to include just the codepoints
used by the list of locales. Reorganized codepoints.yaml to work better
with this script.

Switched from pyyaml to ruamel.yaml for round trip YAML editing.