Refactor and update submodules
Refactored to move submodules and generation scripts into the main
package. After installing the package, generation commands can now be
run with the zmk_locale_generator command.

Settings in pyproject.toml are configured so that only the relevant data
files from the Unicode CLDR and ZMK are included in the installed
package.